move TSelectBuilder to a separate file
